requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
